moved metrics to internal, added proper tags
