+++ modules/rpc/handler_test.go
@@ -25,7 +25,7 @@ import (
 	"errors"
 	"testing"
 
-	"go.uber.org/fx/modules/rpc/stats"
+	"go.uber.org/fx/modules/rpc/internal/stats"
 	"go.uber.org/fx/service"
 	"go.uber.org/thriftrw/wire"
 	"go.uber.org/yarpc/api/transport"
