+++ modules/task/backend.go
@@ -24,7 +24,7 @@ import (
 	"context"
 	"errors"
 
-	"go.uber.org/fx/modules/task/stats"
+	"go.uber.org/fx/modules/task/internal/stats"
 	"go.uber.org/fx/service"
 )
