+++ modules/uhttp/filters.go
@@ -27,7 +27,7 @@ import (
 
 	"go.uber.org/fx"
 	"go.uber.org/fx/auth"
-	"go.uber.org/fx/modules/uhttp/stats"
+	"go.uber.org/fx/modules/uhttp/internal/stats"
 	"go.uber.org/fx/service"
 
 	"github.com/opentracing/opentracing-go"
