+++ modules/uhttp/handler.go
@@ -25,7 +25,7 @@ import (
 	"net/http"
 
 	"go.uber.org/fx"
-	"go.uber.org/fx/modules/uhttp/stats"
+	"go.uber.org/fx/modules/uhttp/internal/stats"
 	"go.uber.org/fx/service"
 )
 
@@ -59,7 +59,7 @@ type handlerWrapper struct {
 // ServeHTTP calls Handler.ServeHTTP(ctx, w, r) and injects a new service context for use.
 func (h *handlerWrapper) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	ctx := fx.NewContext(context.Background(), h.host)
-	stopwatch := stats.HTTPMethodTimer.Timer("http.request." + r.Method).Start()
+	stopwatch := stats.HTTPMethodTimer.Timer(r.Method).Start()
 	defer stopwatch.Stop()
 
 	h.handler.ServeHTTP(ctx, w, r)
