+++ modules/uhttp/internal/stats/metrics.go
@@ -27,30 +27,37 @@ const (
 	TagModule = "module"
 	// TagType is either request or response
 	TagType = "type"
-	// TagProcedure is the procedure name
-	TagProcedure = "procedure"
-	//TagMiddleware is middleware type
+	// TagStatus is status of the request
+	TagStatus = "status"
+	// TagMiddleware is the middleware tag
 	TagMiddleware = "middleware"
 )
 
 // HTTPTags creates metrics scope with defined tags
 var (
-
-	// RPCTags creates metrics scope with defined tags
-	RPCTags = map[string]string{
-		TagModule: "rpc",
+	HTTPTags = map[string]string{
+		TagModule: "http",
 		TagType:   "request",
 	}
 
-	// RPCAuthFailCounter counts auth failures
-	RPCAuthFailCounter tally.Counter
-	// RPCHandleTimer is a turnaround time for rpc handler
-	RPCHandleTimer tally.Scope
+	// HTTPPanicCounter counts panics occurred in http
+	HTTPPanicCounter tally.Counter
+	// HTTPAuthFailCounter counts auth failures
+	HTTPAuthFailCounter tally.Counter
+	// HTTPMethodTimer is a turnaround time for http methods
+	HTTPMethodTimer tally.Scope
+	// HTTPStatusCountScope is a scope for http status
+	HTTPStatusCountScope tally.Scope
 )
 
-// SetupRPCMetrics allocates counters for necessary setup
-func SetupRPCMetrics(scope tally.Scope) {
-	rpcTagsScope := scope.Tagged(RPCTags)
-	RPCAuthFailCounter = rpcTagsScope.Tagged(map[string]string{TagMiddleware: "auth"}).Counter("rpc.auth.fail")
-	RPCHandleTimer = rpcTagsScope.Tagged(RPCTags)
+// SetupHTTPMetrics allocates counters for necessary setup
+func SetupHTTPMetrics(scope tally.Scope) {
+	httpScope := scope.Tagged(HTTPTags)
+	HTTPPanicCounter = httpScope.Counter("panic")
+
+	HTTPAuthFailCounter = httpScope.Tagged(map[string]string{TagMiddleware: "auth"}).Counter("fail")
+
+	HTTPMethodTimer = httpScope.Tagged(HTTPTags)
+
+	HTTPStatusCountScope = httpScope
 }
